move from the EVS API v1.0 to the v1.1
This patch introduces a basic implementation of the API v1.1. The most
of changes are borrowed from the default AOSP implementation.

The ultrasonic feature isn't supported and the related functions aren't
implemented.

The display ownership functions are stubbed (always return success).

Note: this changes must be applied simultaneously for the following
three projects:
 - $ANDROID_BUILD_TOP/vendor/renesas/hal/evs;
 - $ANDROID_BUILD_TOP/vendor/renesas/apps/evs;
 - $ANDROID_BUILD_TOP/device/renesas/common.
This patch breaks support of the API v1.0 (it works only with
IEvsCameraStream of v1.1), that is why it should be applied together
with the according changes for the EVS application.

Test: build, flash and test EVS, ensure that API v1.1 is used:
      # lshal | grep -i evs
Jira-Id: ADM-5092
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>
Change-Id: I37cc6a4511c7058b6b2b72c81fe8e75075e01e85