fill all the required metadata
The EVS API v1.1 provides a set of metadata tags. The HAL doesn't have
enough information to fill all of them by the correct values. So the
current implementation doesn't create metadata for unknown information,
but it appears that for some clients this metadata is essention, and it
is proposed to fill all the required metadata and use zeroes for unknown
values.

From the technical point of view the patch contains the following
changes:
 - add metadata for the following tags:
     - ANDROID_LENS_DISTORTION;
     - ANDROID_LENS_INTRINSIC_CALIBRATION;
     - ANDROID_LENS_POSE_TRANSLATION;
     - ANDROID_LENS_POSE_ROTATION.
 - incapsulate a code for preparing metadata in the separate function;
 - add a metadata retrieving call for the SK+KF boards.

Jira-Id: ADM-5092
Test: try to use the HAL with surround view app.
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>
Change-Id: I59aa680cb40b35a4735c4f2faa825e60fb90235b