explicitly specify hardware environment
The HAL has the same code for all platform. The specific initialization
steps are selected in runtime based on some hardware specific features.
The current version detects the hardware environment based on the
different topology of media nodes. But it appears that on newer BSP
versions this feature can't be no more considered as a board specific.
So it is proposed to explicitly specify the board name where the HAL is
started.

Jira-Id: ADM-5186
Test: launch the HAL on different boards and ensure that it works well.
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>
Change-Id: I877e563eeeed35c53c1d88b1a943afb62f9feca9